+++ packages/e2e-tests/next-app/cypress/integration/redirects.test.ts
@@ -163,7 +163,13 @@ describe("Redirects Tests", () => {
       },
       {
         path: "/query-string-destination-redirect",
-        expectedRedirect: "/ssg-page?a=1234&b=1?",
+        expectedRedirect: "/ssg-page?a=1234&b=1",
+        expectedStatus: 200,
+        expectedRedirectStatus: 308
+      },
+      {
+        path: "/query-string-destination-redirect?foo=bar",
+        expectedRedirect: "/ssg-page?foo=bar&a=1234&b=1",
         expectedStatus: 200,
         expectedRedirectStatus: 308
       }