+++ packages/libs/lambda-at-edge/src/default-handler.ts
@@ -314,7 +314,18 @@ const handleOriginRequest = async ({
     );
     if (customRewrite) {
       rewrittenUri = request.uri;
-      request.uri = customRewrite;
+      const [customRewriteUriPath, customRewriteUriQuery] = customRewrite.split(
+        "?"
+      );
+      request.uri = customRewriteUriPath;
+      if (request.querystring) {
+        request.querystring = `${request.querystring}${
+          customRewriteUriQuery ? `&${customRewriteUriQuery}` : ""
+        }`;
+      } else {
+        request.querystring = `${customRewriteUriQuery ?? ""}`;
+      }
+
       uri = normaliseUri(request.uri);
     }
   }