+++ packages/serverless-components/nextjs-component/src/component.ts
@@ -262,6 +262,8 @@ class NextjsComponent extends Component {
       webACLId: cloudFrontWebACLId,
       restrictions: cloudFrontRestrictions,
       certificate: cloudFrontCertificate,
+      originAccessIdentityId: cloudFrontOriginAccessIdentityId,
+      paths: cloudFrontPaths,
       ...cloudFrontOtherInputs
     } = inputs.cloudfront || {};
 
@@ -621,15 +623,19 @@ class NextjsComponent extends Component {
       comment: cloudFrontComment,
       webACLId: cloudFrontWebACLId,
       restrictions: cloudFrontRestrictions,
-      certificate: cloudFrontCertificate
+      certificate: cloudFrontCertificate,
+      originAccessIdentityId: cloudFrontOriginAccessIdentityId
     });
 
     let appUrl = cloudFrontOutputs.url;
 
-    await createInvalidation({
-      distributionId: cloudFrontOutputs.id,
-      credentials: this.context.credentials.aws
-    });
+    if (!cloudFrontPaths || cloudFrontPaths.length) {
+      await createInvalidation({
+        distributionId: cloudFrontOutputs.id,
+        credentials: this.context.credentials.aws,
+        paths: cloudFrontPaths,
+      });
+    }
 
     const { domain, subdomain } = obtainDomains(inputs.domain);
     if (domain && subdomain) {